fix numba in normalization

+++ seismicpro/gather/utils/normalization.py
@@ -57,12 +57,12 @@ def scale_standard(data, mean, std, tracewise, eps):
     """
     if mean is None and std is None:
         if not tracewise:
-            mean = np.asarray(np.mean(data)).reshape(1, 1)
-            std = np.asarray(np.std(data)).reshape(1, 1)
+            mean = np.asarray(np.mean(data)).reshape(1, 1).astype(data.dtype)
+            std = np.asarray(np.std(data)).reshape(1, 1).astype(data.dtype)
         else:
             trace_len = data.shape[1]
-            mean = np.sum(data, axis=1).reshape(-1, 1) / trace_len
-            std = np.sqrt((np.sum((data - mean)**2, axis=1) / trace_len)).reshape(-1, 1)
+            mean = (np.sum(data, axis=1).reshape(-1, 1) / trace_len).astype(data.dtype)
+            std = np.sqrt((np.sum((data - mean)**2, axis=1) / trace_len)).reshape(-1, 1).astype(data.dtype)
     return (data - mean) / (std + eps)
 
 @njit(nogil=True)
